Move request identity and fetch scheduling types to gosub_net
