+++ crates/gosub_net/Cargo.toml
@@ -21,6 +21,7 @@ thiserror = "1"
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 tokio = { version = "1.52.3", features = ["rt", "io-util", "sync"] }
+tokio-util = { version = "0.7.16", features = ["io"] }
 reqwest = { version = "0.13.3", features = ["rustls", "stream", "blocking"] }
