+++ crates/gosub_net/src/net/mod.rs
@@ -1,5 +1,6 @@
 pub mod events;
 pub mod null_emitter;
 pub mod observer;
+pub mod request_ref;
 pub mod types;
 pub mod utils;
